Moved Parent Processor to new file
